Move v4 functions into `task-list-generator.js`
