+++ packages/webapp/src/journey-validation/task-list-generator.js
@@ -107,7 +107,7 @@ const getTaskStatus = (task, session) => {
   }
 }
 
-const getTaskItemsV5 = (task, session) => {
+const getTaskItems = (task, session) => {
   const calculatedStatus = checkTaskStatus(task, session)
   return {
     id: task.id,
@@ -130,27 +130,27 @@ const retrieveTask = (routeDefinitions, startUrl) => {
   return routeDefinitions.find(route => route.startUrl === startUrl) || null
 }
 
-const generateTaskListV5 = (taskSections, session) => {
+const generateTaskList = (taskSections, session) => {
   const taskList = taskSections.map(section => ({
     taskTitle: section.title,
-    items: section.tasks.map(task => getTaskItemsV5(task, session))
+    items: section.tasks.map(task => getTaskItems(task, session))
   }))
   return taskList
 }
 
-const statusForDisplayV5 = status => status.charAt(0).toUpperCase() + status.slice(1).toLowerCase()
+const statusForDisplay = status => status.charAt(0).toUpperCase() + status.slice(1).toLowerCase()
 
 // Recursively iterate over a task list object and convert each `text` property (which will be the task item status) to
 // have an initial capital with the rest of the string lower case.
-const formatStatusesForDisplayV5 = taskList => {
+const formatStatusesForDisplay = taskList => {
   if (Array.isArray(taskList)) {
-    taskList.forEach(item => formatStatusesForDisplayV5(item))
+    taskList.forEach(item => formatStatusesForDisplay(item))
   } else if (typeof taskList === 'object' && taskList !== null) {
     Object.keys(taskList).forEach(key => {
       if (key === 'text' && typeof taskList[key] === 'string') {
-        taskList[key] = statusForDisplayV5(taskList[key])
+        taskList[key] = statusForDisplay(taskList[key])
       } else {
-        formatStatusesForDisplayV5(taskList[key])
+        formatStatusesForDisplay(taskList[key])
       }
     })
   }
@@ -163,19 +163,19 @@ const getTaskList = (journey, session) => {
   // prevent the original version being affected by this.
   switch (journey) {
     case constants.applicationTypes.REGISTRATION:
-      taskList = generateTaskListV5(registrationTaskSections, session)
+      taskList = generateTaskList(registrationTaskSections, session)
       taskList.push(structuredClone(registrationCheckYourAnswers))
       break
     case constants.applicationTypes.CREDITS_PURCHASE:
-      taskList = generateTaskListV5(creditsPurchaseTaskSections, session)
+      taskList = generateTaskList(creditsPurchaseTaskSections, session)
       taskList.push(structuredClone(creditsPurchaseCheckYourAnswers))
       break
     case constants.applicationTypes.ALLOCATION:
-      taskList = generateTaskListV5(allocationTaskSections, session)
+      taskList = generateTaskList(allocationTaskSections, session)
       taskList.push(structuredClone(allocationCheckYourAnswers))
       break
     case constants.applicationTypes.COMBINED_CASE:
-      taskList = generateTaskListV5(combinedCaseTaskSections, session)
+      taskList = generateTaskList(combinedCaseTaskSections, session)
       taskList.push(combinedCaseCheckYourAnswers)
       break
     default:
@@ -215,7 +215,7 @@ const getTaskList = (journey, session) => {
   // v5 gov uk frontend has statuses displayed in lower case with an initial capital rather than all caps as before. We
   // may be reliant elsewhere on the status being all caps, so for now instead of changing the statuses to be the
   // correct case throughout the code, we simply reformat them at this point once we're ready to display them
-  formatStatusesForDisplayV5(taskList)
+  formatStatusesForDisplay(taskList)
 
   return { taskList, totalTasks, completedTasks, canSubmit }
 }
@@ -258,8 +258,101 @@ const getNextStep = (request, h, errCallback) => {
   throw new Error('Next URL is not set')
 }
 
+const generateTaskListV4 = (taskSections, session) => {
+  const locked = (section, taskList) => {
+    if (section.dependantIds && section.dependantIds.length > 0) {
+      for (const dependantId of section.dependantIds) {
+        const dependantSection = taskList.find(s => s.id === dependantId)
+        if (dependantSection) {
+          const uncompletedTasks = dependantSection.tasks.filter(task => task.status !== constants.COMPLETE_REGISTRATION_TASK_STATUS)
+          if (uncompletedTasks.length > 0) {
+            return true
+          }
+        }
+      }
+    }
+    return false
+  }
+  const taskList = taskSections.map(section => {
+    return {
+      taskTitle: section.title,
+      tasks: section.tasks.map(task => getTaskStatus(task, session)),
+      id: section.id,
+      dependantIds: section.dependantIds
+    }
+  })
+
+  const lockedTaskList = taskList.map(section => {
+    const isLocked = locked(section, taskList)
+    if (isLocked) {
+      return {
+        ...section,
+        ...{
+          tasks: section.tasks.map(task => {
+            task.status = constants.CANNOT_START_YET_STATUS
+            task.isLocked = true
+            return task
+          })
+        }
+      }
+    }
+    return section
+  })
+
+  return lockedTaskList
+}
+
+const getTaskListV4 = (journey, session) => {
+  let taskList
+
+  switch (journey) {
+    case constants.applicationTypes.REGISTRATION:
+      taskList = generateTaskListV4(registrationTaskSections, session)
+      taskList.push(registrationCheckYourAnswers)
+      break
+    case constants.applicationTypes.CREDITS_PURCHASE:
+      taskList = generateTaskListV4(creditsPurchaseTaskSections, session)
+      taskList.push(creditsPurchaseCheckYourAnswers)
+      break
+    case constants.applicationTypes.ALLOCATION:
+      taskList = generateTaskListV4(allocationTaskSections, session)
+      taskList.push(allocationCheckYourAnswers)
+      break
+    case constants.applicationTypes.COMBINED_CASE:
+      taskList = generateTaskListV4(combinedCaseTaskSections, session)
+      taskList.push(combinedCaseCheckYourAnswers)
+      break
+    default:
+      taskList = []
+  }
+
+  let completedTasks = 0
+  let totalTasks = 0
+
+  taskList.forEach(task => {
+    if (task.tasks.length === 1) {
+      totalTasks += 1
+      if (task.tasks[0].status === constants.COMPLETE_REGISTRATION_TASK_STATUS) {
+        completedTasks += 1
+      }
+    } else {
+      task.tasks.forEach(currentTask => {
+        totalTasks += 1
+        if (currentTask.status === constants.COMPLETE_REGISTRATION_TASK_STATUS) {
+          completedTasks += 1
+        }
+      })
+    }
+  })
+
+  const canSubmit = completedTasks === (totalTasks - 1)
+
+  return { taskList, totalTasks, completedTasks, canSubmit }
+}
+
 export {
   getTaskList,
+  getTaskListV4,
   getIndividualTaskStatus,
   getNextStep
 }
