+++ packages/webapp/src/routes/combined-case/tasklist.js
@@ -1,5 +1,5 @@
 import constants from '../../utils/constants.js'
-import { getTaskList } from '../../journey-validation/task-list-generator-v4.js'
+import { getTaskListV4 as getTaskList } from '../../journey-validation/task-list-generator.js'
 
 const handlers = {
   get: async (request, h) => {
